# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: HCML Monitoring Portal (BFF)
  version: 1.0.0
  summary: A read-only backend-for-frontend over the Zabbix 7.0 JSON-RPC API.
  description: |
    The portal owns **no database**. Every response below is either passed through from Zabbix or
    derived in-process and cached in a 500-entry LRU that is lost on restart. Sites, links,
    incidents, host reachability and the monthly SLA have no table anywhere.

    **One write exists** in the whole surface: `POST /api/problems/acknowledge`. It uses a separate
    write-capable Zabbix token and is refused with 503 when that token is unset.

    ### Authentication

    Two schemes, and both work on every guarded path:

    - `Authorization: Bearer <jwt>`, the normal one, from `POST /api/auth/login`. 12-hour expiry.
    - `?token=<jwt>`: a query-string alternative. It exists because `EventSource` cannot set a
      header, but `auth.ts:129-132` reads it on **every** `/api/*` request, not only the SSE ones.
      `logging.ts` redacts it from the request log.

    When `AUTH_ENABLED=false` no route is guarded at all and every caller is treated as `admin`.

    ### Roles

    OpenAPI cannot express a role requirement per path, so each operation states its role in the
    description. From `ROUTE_RULES` (`server/src/auth.ts:35-40`), first match wins, unmatched is
    `viewer`:

    | Role | Paths |
    |---|---|
    | `admin` | `/api/reports/inventory` |
...
    | `viewer` | everything else |

    ### Errors

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/history:
    get:
      summary: Raw history rows for one item
      description: |
        Cached 15 s. The window ends at the item's newest stored clock, not at wall clock, so a stalled
        item still returns its last real data rather than an empty range.

        **No row limit is applied.** At `hours=168` on a 60 s item that is 10,080 rows.
      operationId: getHistory
      tags: [History]
      parameters:
        - name: itemid
          in: query
          required: true
          schema: { type: string, pattern: '^\d+$' }
        - name: hours
          in: query
          schema: { type: integer, minimum: 1, maximum: 168, default: 1 }
        - name: history
          in: query
          description: Zabbix `value_type`, 0 float, 1 char, 2 log, 3 unsigned, 4 text.
          schema: { type: integer, minimum: 0, maximum: 4, default: 0 }
      responses:
        '200':
          description: Rows as Zabbix returns them.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    itemid: { type: string }
                    clock: { type: string }
                    value: { type: string }
                  additionalProperties: true
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/trend:
    get:
      summary: Hourly trend rows for one item
      description: |
        Cached 60 s. Trends are Zabbix's hourly min/avg/max aggregate and reach back about 2.2 years
        in this dataset, against 30–90 days for raw history.
      operationId: getTrend
      tags: [History]
      parameters:
        - name: itemid
          in: query
          required: true
          schema: { type: string, pattern: '^\d+$' }
        - name: hours
          in: query
          schema: { type: integer, minimum: 1, maximum: 8760, default: 168 }
      responses:
        '200':
          description: Hourly aggregates.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    itemid: { type: string }
                    clock: { type: string }
                    num: { type: string }
                    value_min: { type: string }
                    value_avg: { type: string }
                    value_max: { type: string }
                  additionalProperties: true
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/graph:
    get:
      summary: Downsampled series for up to four items
      description: |
        Cached 15 s live, 600 s historical. **Three mutually exclusive range modes, in precedence
        order: `month`, then `from`+`to`, then `hours`.**

        Below 24 hours it reads raw history; above, hourly trends; and `source` in the response says
        which was used. Everything is bucketed to at most `points` per series, so a 30-day graph is a
        few tens of kB rather than 500. An explicit `null` point is inserted wherever the gap between
        samples exceeds 3× the expected spacing, so the chart breaks the line instead of drawing
        across a collection gap.
      operationId: getGraph
      tags: [History]
      parameters:
        - name: itemids
          in: query
          required: true
          description: Comma-separated. **At most 4** or 400.
          schema: { type: string }
        - name: month
          in: query
          description: Highest precedence. A future month is 400.
          schema: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        - name: from
          in: query
          description: Unix seconds. Must be paired with `to`. Range may not exceed 400 days.
          schema: { type: integer }
        - name: to
          in: query
          schema: { type: integer }
        - name: hours
          in: query
          description: Used only when neither `month` nor `from`/`to` is given.
          schema: { type: integer, minimum: 1, maximum: 8760, default: 24 }
        - name: points
          in: query
          description: Target samples per series after bucketing.
          schema: { type: integer, minimum: 200, maximum: 3000, default: 1500 }
      responses:
        '200':
          description: One series per item.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GraphResponse' }
        '400':
          description: |
            Bad range, more than 4 items, unknown itemids, or a non-numeric item.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The normal scheme. 12-hour expiry, carries `{ sub, role }`.
    queryToken:
      type: apiKey
      in: query
      name: token
      description: |
        The same JWT in the query string. Required for `EventSource`, which cannot set headers, but
        accepted on **every** guarded path. Kept out of the request log by `logging.ts`.
  responses:
    BadRequest:
      description: '`bad_request`, a parameter failed validation.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: '`unauthorized`, missing or expired portal JWT. Not the same as `zabbix_auth`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ZabbixError:
      description: '`zabbix_error`, unreachable, non-2xx, non-JSON-RPC, or a JSON-RPC error.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ZabbixUnavailable:
      description: |
        `zabbix_timeout` (no answer within `ZABBIX_TIMEOUT_MS`, default 10 s) or `zabbix_auth` (Zabbix
        rejected the API token). Both are configuration states rather than upstream faults, which is
        why they are 503 and not 504.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error, message]
      description: The only error shape this server produces, at every status.
      properties:
        error: { type: string, description: A stable machine-readable code. }
        message: { type: string, description: Safe to show a user. Never contains the Zabbix URL. }
...
    GraphResponse:
      type: object
      required: [series, from, to, source]
      properties:
        from: { type: integer }
        to: { type: integer }
        source:
          type: string
          enum: [history, trend, 'trend+history']
          description: Which store answered. `trend+history` means trends stalled and raw filled the tail.
        series:
          type: array
          items:
            type: object
            required: [itemid, name, points]
            properties:
              itemid: { type: string }
              name: { type: string }
              units: { type: string }
              host: { type: string }
              step: { type: boolean, description: Render as a step line, ICMP and other 0/1 series. }
              points:
                type: array
                description: '`[clock, avg, min, max]`. A null `avg` is an explicit collection gap.'
                items:
                  type: array
                  items:
                    oneOf: [{ type: number }, { type: 'null' }]
              stats:
                oneOf:
                  - type: object
                    additionalProperties: true
                  - type: 'null'
              coverage: { type: number }
...